+++ src/Krystal/Date/TimeHelper.php
@@ -11,6 +11,8 @@
 
 namespace Krystal\Date;
 
+use LogicException;
+
 /**
  * Time related helper methods 
  */
@@ -31,6 +33,29 @@ abstract class TimeHelper
     public static function getQuarters()
     {
         return range(1, 4);
+    }
+
+    /**
+     * Returns a collection of months by associated quarter
+     * 
+     * @param integer $quarter
+     * @throws \LogicException If invalid quarter supplied
+     * @return array
+     */
+    public static function getMonthsByQuarter($quarter)
+    {
+        switch ($quarter) {
+            case 1:
+                return array('01', '02', '03');
+            case 2:
+                return array('04', '05', '06');
+            case 3:
+                return array('07', '08', '09');
+            case 4:
+                return array('10', '11', '12');
+            default:
+                throw new LogicException(sprintf('Invalid quarter supplied - %s', $quarter));
+        }
     }
 
     /**